requirements.txt
Flask==2.0.1
tensorflow==2.15.0
werkzeug==2.0.1
Pillow==8.2.0
flask_sqlalchemy
flask_mail
gunicorn
